+++ src/core/keymap/index.ts
@@ -5,4 +5,6 @@ export * from './validator';
 export * from './oled';
 export * from './wpm';
 export * from './layers';
+export * from './generator';
+export * from './converter';
 export type { Keymap, Layer, KeymapConfig } from '../../types';